11-Inner_Working_Function_Tools/pyproject.toml
[project]
name = "11-inner-working-function-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
